Deduplicate WAT orchestration helpers
