+++ tools/watwat.test.mjs
@@ -1,4 +1,3 @@
-import fs from "node:fs/promises";
 import { createRequire } from "node:module";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
@@ -10,12 +9,12 @@ import {
 
 const require = createRequire(import.meta.url);
 const {
-  ASSERT_FAILURE_EXPECTED_MESSAGE,
-  assertFailureProbeExportNames,
+  assertFailureProbeCases,
+  findTestWasms,
 } = require("./watwat-core.js");
 
 const ROOT_DIR = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
-const WAT_TEST_DIRS = Object.freeze(["dist/wasm", "dist/wasm/std"]);
+const WAT_TEST_ROOT = "dist/wasm";
 const ASSERT_FAILURE_PROBE_FILE = "dist/wasm/std/assert.test.wasm";
 const EXPECTED_FAILURE_PROBE_FILES = Object.freeze([
   "dist/wasm/watwat.test.wasm",
@@ -26,27 +25,9 @@ function normalizedWatTestPath(file) {
   return file.replaceAll(path.sep, "/");
 }
 
-async function watTestFilesIn(dir) {
-  let entries;
-
-  try {
-    entries = await fs.readdir(path.join(ROOT_DIR, dir), { withFileTypes: true });
-  } catch (error) {
-    if (error.code === "ENOENT") {
-      return [];
-    }
-
-    throw error;
-  }
-
-  return entries
-    .filter((entry) => entry.isFile() && entry.name.endsWith(".test.wasm"))
-    .map((entry) => normalizedWatTestPath(path.join(dir, entry.name)));
-}
-
 const watTestFiles = (
-  await Promise.all(WAT_TEST_DIRS.map((dir) => watTestFilesIn(dir)))
-).flat().sort();
+  await findTestWasms(path.join(ROOT_DIR, WAT_TEST_ROOT), { allowMissing: true })
+).map((file) => normalizedWatTestPath(path.relative(ROOT_DIR, file)));
 const watTestFileSet = new Set(watTestFiles);
 const missingExpectedFailureProbeFiles = EXPECTED_FAILURE_PROBE_FILES.filter(
   (file) => !watTestFileSet.has(normalizedWatTestPath(file)),
@@ -70,10 +51,9 @@ const expectedFailureProbes = [
     expectedMessage: "deliberate i32 failure",
     file: "dist/wasm/watwat.test.wasm",
   },
-  ...(await assertFailureProbeExportNames(path.join(ROOT_DIR, ASSERT_FAILURE_PROBE_FILE))).map(
-    (exportName) => ({
-      exportName,
-      expectedMessage: ASSERT_FAILURE_EXPECTED_MESSAGE,
+  ...(await assertFailureProbeCases(path.join(ROOT_DIR, ASSERT_FAILURE_PROBE_FILE))).map(
+    (probe) => ({
+      ...probe,
       file: ASSERT_FAILURE_PROBE_FILE,
     }),
   ),
